requirements.txt
torch
transformers
pillow
decord
opencv-python
